Remove thread local checks on mod container setting, returning Forge behaviour.
This reversion is happening after discussions with @bloodmc on performance.

Effectively, if mods are using LoadController#activeContainer(), it's on them - from comments in Forge:

//Instructs event handlers to set the active mod container.
//This has a major performance impact so use sparingly.
//Note: The context IS NOT thread aware as this would be ungodly slow.
//So This should ONLY be used for Forge mod lifecycle events.

We were kind of breaking that contract by doing this, but it allowed us
to fix issues with mods that were doing it. They shouldn't be now.

Plus this should give a nice performance boost.

Reverts "Make calls to the active mod container thread local."
This reverts commit 0b92ea4.